feat: Add PQC mode config, decoderawtransaction PQC display, and BearTec wallet integration library
- Add -pqcmode=hybrid|classical|pure config option in init.cpp
- Add PQCMode enum to pqc_config.h with safe defaults (PQC off by default)
- Update decoderawtransaction to show PQC metadata (algorithm, mode, sizes, sig/pubkey hex)
- Improve PQC signing logging (LogDebug for per-input, LogPrintf for errors)
- Fix -Wsign-compare warning in hybrid_key.cpp
- Add BearTec wallet integration library (contrib/beartec-wallet/):
  - Dilithium key derivation matching node's HMAC-SHA512 scheme
  - Bech32 qbtct1/qbtc1 address generation
  - 2-of-3 Shamir secret sharing for hybrid PQC keys
  - JSON-RPC client with cookie auth
  - Full self-test suite

+++ src/crypto/pqc/hybrid_key.cpp
@@ -110,7 +110,7 @@ bool HybridKey::Verify(const uint256& hash, const std::vector<unsigned char>& si
     if (signature.size() < 2) return false;
     
     uint8_t csig_len = signature[0];
-    if (signature.size() < 1 + csig_len + 1) return false;
+    if (signature.size() < static_cast<size_t>(1 + csig_len + 1)) return false;
     
     std::vector<unsigned char> classical_sig(signature.begin() + 1, 
                                             signature.begin() + 1 + csig_len);
